Add number_sequences table and nullable 'number' column on areas and iterations tables (#2309)
This adds back the `number_sequences` table the `number` columns on the `areas` and `iterations` tables (known from #2287 but then reverted). But it does this in three individual steps. This change is backwards compatible with the old code as the `number` column is nullable. Except for this structural change, no data is changed.